github-app/pyproject.toml
[project]
name = "minis-github-app"
version = "0.1.0"
description = "GitHub App for PR reviews by minis (AI personality clones)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.4",
    "fastapi>=0.128.6",
    "httpx>=0.28.1",
    "litellm>=1.81.9",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "uvicorn>=0.40.0",
]
